first commit, forgot to create a branch before i started work refactoring so its all in one commit
