+++ src/consts.js
@@ -1,0 +1,6 @@
+
+const LOAD_BLOCKS = 50;
+
+module.exports = {
+	LOAD_BLOCKS
+}
