+++ src/test/shared.js
@@ -1,0 +1,9 @@
+
+const f = (name) => {
+	const thing = name;
+
+	return thing
+}
+
+console.log(f("1"))
+console.log(f("2"))
